fix: vocatius, and cleanup wayland stuff
well linux as my primary system will probably never happen, at least not
in the pace as it is progressing right now.
The whole linux desktop experience is a total mess i don't really
believe that wayland is going anywhere.
The protocol seems to be "poorly" designed, the "security" measures
could have been implemented in a way that is non breaking, and
considering there's basically only wlroots is a testament to how complex
this protocol is. yes gnome and kde are a thing, but they're tailored
for its specific usecases.

Let's pray that arcan will be the way to move forward.